Build using hacked tech-docs gem
Before running `make build` I added some hacks to
my local copies of the gem. The main one is to 
rewrite the "url" which appears in the search.json
file, to include the docpath prefix of the site 
(in this case "/technical-guidance").

This is to see if I can get search to work on a 
site which is deployed using github pages, and not
running at the root of its domain (normal github
pages sites run at /[repo name] by default).

If this hack works, I'll try and find a way to do
it which is a bit less hacky.